+++ test/test_stabs.jl
@@ -46,6 +46,9 @@ function test_stabs()
             canonicalize!(s)
             dss = canonicalize!(copy(stabilizerview(ds)))
             @test s == dss
+            stabs = [s[1:i] for s in [random_stabilizer(n) for n in [32,16,16,64,63,65,129,128,127]] for i in rand(1:10)];
+            mdstabs = MixedDestabilizer.(stabs);
+            @test canonicalize!(⊗(stabs...)) == canonicalize!(stabilizerview(⊗(mdstabs...)))
         end
     end
     @testset "Stabilizer indexing" begin