+++ NEWS.md
@@ -1,9 +1,11 @@
 # News
 
-## master
+## v0.3.0
 
-- Support for graph states, e.g., conversion to and from arbitrary `Stabilizer` state.
+- `dot` and `logdot` for calculating dot products between `Stabilizer` states.
+- Initial support for graph states, e.g., conversion to and from arbitrary `Stabilizer` state.
 - Implemented `Makie.jl` plotting recipes in the `QuantumCliffordPlots.jl` package, which now also stores the `Plots.jl` recipes.
+- Much faster `tensor` product of states.
 - `CliffordColumnForm` has been completely removed. Only `CliffordOperator` now exists.
 - `random_singlequbitop` was removed, as it was using `CliffordColumnForm`. `random_clifford1` is a partial alternative.
 - Drop `Require` to improve import times.